+++ src/dpe_base/io/io_handler.cc
@@ -55,7 +55,7 @@ bool IOHandler::Read()
   BOOL fSuccess = ReadFile(
                io_handle_,
                (char*)read_buffer_.c_str(),
-               read_buffer_.size(),
+               static_cast<int>(read_buffer_.size()),
                (DWORD*)&read_size_,
                ol);
